requirements.txt
# Core web framework
fastapi==0.115.0
uvicorn[standard]==0.30.5

# Data models & validation
pydantic==2.9.0

# Logging
structlog==24.2.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9   # Only if using PostgreSQL; safe to remove if SQLite-only

# Data analysis
pandas==2.2.3
openpyxl==3.1.5

# Background job & queue
redis==5.0.8
rq==1.16.2

# PPTX generation
python-pptx==0.6.23

# HTTP requests
requests==2.32.3

# Optional dev tools
python-dotenv==1.0.1

# Security
passlib[bcrypt]==1.7.4
python-jose==3.3.0
PyJWT==2.9.0

pydantic==2.9.0
pydantic-settings==2.6.1

email-validator==2.2.0